+++ app/Changelog.md
@@ -1,10 +1,10 @@
 # Changelog
 
 ## Version 0.5.0
-- Add budgeting of a wallet
+- Add budgeting of a wallet.
 
 ## Version 0.4.5
 - Bug fix in the page AllTransactions when one transaction date is the first day of a month.
 
 ## Version 0.4.2
-- Add new category in ManageCategories through an AlertDialog
+- Add new category in ManageCategories through an AlertDialog.